enpoints related to vertex metadata addition and deletion are disabled if the confi file commands so (to prevent users from editing the l ables using the cmd and by-passing the FE)

+++ be/server/vertex_metadata/controller.py
@@ -8,6 +8,7 @@ from . import VertexMetadataSchema
 from .model import VertexMetadata
 from .service import VertexMetadataService
 from .. import SessionLocal
+from ...configuration import CONFIGURATIONS
 
 api = Namespace("VertexMetadata", description="Global metadata related to eth addresses")
 
@@ -27,6 +28,9 @@ class MetadataResource(Resource):
     @accepts(schema=VertexMetadataSchema, api=api)
     @responds(schema=VertexMetadataSchema)
     def post(self) -> VertexMetadata:
+        if CONFIGURATIONS['is_labelling_enabled'] != 'true':
+            print("\tLabelling is disabled in config, this endpoint is not active")
+            return {}
         with SessionLocal() as db:
             by_eth = VertexMetadataService.create(request.parsed_obj, db)
             return by_eth
@@ -37,6 +41,9 @@ class MetadataResource(Resource):
 
     @responds(schema=VertexMetadataSchema)
     def delete(self, eth: str, typee: str, value: str) -> VertexMetadata:
+        if CONFIGURATIONS['is_labelling_enabled'] != 'true':
+            print("\tLabelling is disabled in config, this endpoint is not active")
+            return {}
         with SessionLocal() as db:
             VertexMetadataService.delete(eth, typee, value, db)
             return {}
